relative position from top to bottom

+++ internal/duck_dom/border.go
@@ -90,8 +90,8 @@ func render_border(position Position, styles *Styles) string {
 	border_builder.WriteString(top)
 
 	for i := 1; uint(i) < uint(styles.Height)-1; i += 1 {
-		left_wall := fmt.Sprintf(MOVE_CURSOR_TO_POSITION, position.StartingRow+uint(i), position.StartingCol)
-		right_wall := fmt.Sprintf(MOVE_CURSOR_TO_POSITION, position.StartingRow+uint(i), position.StartingCol+uint(styles.Width)-1)
+		left_wall := fmt.Sprintf(MOVE_CURSOR_TO_POSITION, position.StartingRow+i, position.StartingCol)
+		right_wall := fmt.Sprintf(MOVE_CURSOR_TO_POSITION, position.StartingRow+i, position.StartingCol+styles.Width-1)
 		wall := left_wall + styles.Border.Style.Left + right_wall + styles.Border.Style.Right
 		border_builder.WriteString(wall)
 	}
